Enhancement : Make CommandProcessorImpl Autoclosable, so the scanner would be closed in case of exception
