+++ src/main/java/Dude.java
@@ -1,9 +1,9 @@
-import commandImpl.CommandLogicUnitImpl;
-import commandImpl.CommandProcessorImpl;
-import commandInterface.ICommandLogicUnit;
-import commandInterface.ICommandProcessor;
+import commandimpl.CommandLogicUnitImpl;
+import commandimpl.CommandProcessorImpl;
 import dao.TaskDao;
 import dao.TaskDaoImpl;
+import icommand.ICommandLogicUnit;
+import icommand.ICommandProcessor;
 
 import static util.Display.printSentence;
 
@@ -14,7 +14,6 @@ public class Dude {
 	// initialize the command processor from logic processing, use the commandParse to process the console input
 	private static final TaskDao taskDao = new TaskDaoImpl();
 	private static final ICommandLogicUnit commandLogicUnit = new CommandLogicUnitImpl(taskDao);
-	private static final ICommandProcessor commandProcessor = new CommandProcessorImpl(commandLogicUnit);
 	
 	@SuppressWarnings("InfiniteLoopStatement")
 	public static void main(String[] args) {
@@ -27,8 +26,12 @@ public class Dude {
 		
 		starting();
 		
-		while (true) {
-			commandProcessor.processInput();
+		try (ICommandProcessor commandProcessor = new CommandProcessorImpl(commandLogicUnit)) {
+			while (true) {
+				commandProcessor.processInput();
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
 		}
 	}
